Unify configs
